requirements-deploy.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
greenlet==3.0.3
pydantic[email]==2.7.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
qrcode[pil]==7.4.2
Pillow==10.3.0
cryptography==42.0.5
pdf2image==1.17.0
gdown==5.1.0
web3==6.15.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.0+cpu
torchvision==0.18.0+cpu